refactor: rename VoiceState.py -> active_voice_state.py
